# apps/backend/docs/swagger.yaml
basePath: /
definitions:
...
  requests.UpdatePasswordInput:
    properties:
      current_password:
        maxLength: 255
        minLength: 8
        type: string
      new_password:
        maxLength: 255
        minLength: 8
        type: string
      new_password_confirmation:
        type: string
    required:
    - current_password
    - new_password
    - new_password_confirmation
    type: object
  requests.UpdateProfileInput:
    properties:
      email:
        type: string
      name:
        type: string
      phone:
        type: string
    type: object
...
  responses.Pagination:
    properties:
      from:
        type: integer
      has_more:
        type: boolean
      page:
        type: integer
      page_size:
        type: integer
      to:
        type: integer
      total_items:
        type: integer
      total_pages:
        type: integer
    type: object
  utils.OrderedErrorResponse:
    properties:
      details: {}
      dev_message:
        type: string
      error_code:
        type: string
      errors: {}
      message:
        type: string
      success:
        type: boolean
      trace:
        type: string
    type: object
  utils.OrderedSuccessResponse:
    properties:
      data: {}
      message:
        type: string
      meta: {}
      pagination:
        $ref: '#/definitions/responses.Pagination'
      success:
        type: boolean
    type: object
host: localhost:9001
info:
  contact: {}
  description: Ini adalah dokumentasi API untuk backend berbasis Go Fiber.
  title: DilZ Backend API
  version: "1.0"
paths:
...
  /v1/payment-gateways:
    get:
      consumes:
      - application/json
      description: Mengambil semua data konfigurasi payment gateway yang tersedia
        di sistem.
      produces:
      - application/json
      responses:
        "200":
          description: Berhasil mendapatkan data payment gateway
          schema:
            $ref: '#/definitions/utils.OrderedSuccessResponse'
      summary: Ambil daftar payment gateway
      tags:
      - Payment Gateways
    post:
      consumes:
      - application/json
      description: Membuat entri data konfigurasi payment gateway baru ke dalam database.
      parameters:
      - description: Data JSON payment gateway baru
        in: body
        name: payload
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: Berhasil menambahkan payment gateway
          schema:
            $ref: '#/definitions/utils.OrderedSuccessResponse'
      summary: Tambah payment gateway baru
      tags:
      - Payment Gateways
  /v1/payment-gateways/{id}:
    delete:
      consumes:
      - application/json
      description: Menghapus data konfigurasi payment gateway dari database berdasarkan
        ID.
      parameters:
      - description: ID Payment Gateway
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Berhasil menghapus payment gateway
          schema:
            $ref: '#/definitions/utils.OrderedSuccessResponse'
      summary: Hapus payment gateway
      tags:
      - Payment Gateways
    get:
      consumes:
      - application/json
      description: Mengambil satu baris data konfigurasi payment gateway secara spesifik
        menggunakan parameter ID.
      parameters:
      - description: ID Payment Gateway
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Berhasil mendapatkan data payment gateway
          schema:
            $ref: '#/definitions/utils.OrderedSuccessResponse'
      summary: Ambil detail payment gateway
      tags:
      - Payment Gateways
    put:
      consumes:
      - application/json
      description: Memodifikasi keseluruhan data konfigurasi payment gateway berdasarkan
        ID.
      parameters:
      - description: ID Payment Gateway
        in: path
        name: id
        required: true
        type: string
      - description: Data JSON pembaruan payment gateway
        in: body
        name: payload
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Berhasil mengupdate payment gateway
          schema:
            $ref: '#/definitions/utils.OrderedSuccessResponse'
      summary: Update payment gateway
      tags:
      - Payment Gateways
...
  /v1/settings:
    get:
      consumes:
      - application/json
      description: Mengambil semua data konfigurasi dan pengaturan aplikasi.
      produces:
      - application/json
      responses:
        "200":
          description: Berhasil mendapatkan data setting
          schema:
            $ref: '#/definitions/utils.OrderedSuccessResponse'
      summary: Ambil daftar pengaturan
      tags:
      - Settings
    post:
      consumes:
      - application/json
      description: Membuat data konfigurasi pengaturan baru ke dalam database.
      parameters:
      - description: Data JSON pengaturan baru
        in: body
        name: payload
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: Berhasil menambahkan setting
          schema:
            $ref: '#/definitions/utils.OrderedSuccessResponse'
      summary: Tambah pengaturan baru
      tags:
      - Settings
  /v1/settings/{id}:
    delete:
      consumes:
      - application/json
      description: Menghapus data pengaturan dari database berdasarkan ID.
      parameters:
      - description: ID pengaturan
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Berhasil menghapus setting
          schema:
            $ref: '#/definitions/utils.OrderedSuccessResponse'
      summary: Hapus pengaturan
      tags:
      - Settings
    put:
      consumes:
      - application/json
      description: Memodifikasi data pengaturan berdasarkan ID.
      parameters:
      - description: ID pengaturan
        in: path
        name: id
        required: true
        type: string
      - description: Data JSON pembaruan pengaturan
        in: body
        name: payload
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Berhasil mengupdate setting
          schema:
            $ref: '#/definitions/utils.OrderedSuccessResponse'
      summary: Update pengaturan
      tags:
      - Settings
  /v1/settings/{key}:
    get:
      consumes:
      - application/json
      description: Mengambil nilai konfigurasi atau pengaturan spesifik berdasarkan
        'key' (kunci) pengaturannya.
      parameters:
      - description: 'Key string dari pengaturan (contoh: ''site_title'')'
        in: path
        name: key
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Berhasil mendapatkan data setting
          schema:
            $ref: '#/definitions/utils.OrderedSuccessResponse'
        "500":
          description: Gagal mendapatkan data setting
          schema:
            $ref: '#/definitions/utils.OrderedErrorResponse'
      summary: Ambil detail pengaturan
      tags:
      - Settings
  /v1/users:
    get:
      consumes:
      - application/json
      description: Mengambil semua data pengguna dari database. (Membutuhkan hak akses
        admin).
      produces:
      - application/json
      responses:
        "200":
          description: User berhasil diambil
          schema:
            $ref: '#/definitions/utils.OrderedSuccessResponse'
        "500":
          description: Gagal mengambil user
          schema:
            $ref: '#/definitions/utils.OrderedErrorResponse'
      security:
      - BearerAuth: []
      summary: Ambil daftar user
      tags:
      - Users
  /v1/users/{id}:
    get:
      consumes:
      - application/json
      description: Mengambil satu data pengguna secara spesifik menggunakan parameter
        ID. (Membutuhkan hak akses admin).
      parameters:
      - description: ID User
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: User berhasil diambil
          schema:
            $ref: '#/definitions/utils.OrderedSuccessResponse'
        "404":
          description: User tidak ditemukan
          schema:
            $ref: '#/definitions/utils.OrderedErrorResponse'
      security:
      - BearerAuth: []
      summary: Ambil detail user
      tags:
      - Users
  /v1/users/password:
    patch:
      consumes:
      - application/json
      description: Memperbarui kata sandi pengguna berdasarkan token JWT yang sedang
        aktif.
      parameters:
      - description: Data kata sandi lama dan baru
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/requests.UpdatePasswordInput'
      produces:
      - application/json
      responses:
        "200":
          description: Password berhasil diperbarui
          schema:
            $ref: '#/definitions/utils.OrderedSuccessResponse'
        "401":
          description: Error validasi form atau password lama salah
          schema:
            $ref: '#/definitions/utils.OrderedErrorResponse'
        "500":
          description: Terjadi kesalahan server
          schema:
            $ref: '#/definitions/utils.OrderedErrorResponse'
      security:
      - BearerAuth: []
      summary: Update password user
      tags:
      - Users
  /v1/users/profile:
    patch:
      consumes:
      - application/json
      description: Memperbarui data profil pengguna berdasarkan token JWT yang sedang
        aktif.
      parameters:
      - description: Data profil yang ingin diubah
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/requests.UpdateProfileInput'
      produces:
      - application/json
      responses:
        "200":
          description: Profile berhasil diperbarui
          schema:
            $ref: '#/definitions/utils.OrderedSuccessResponse'
        "401":
          description: Error validasi form input
          schema:
            $ref: '#/definitions/utils.OrderedErrorResponse'
        "500":
          description: Terjadi kesalahan server
          schema:
            $ref: '#/definitions/utils.OrderedErrorResponse'
      security:
      - BearerAuth: []
      summary: Update profil user
      tags:
      - Users
swagger: "2.0"